+++ src/wininput.c
@@ -1915,7 +1915,7 @@ static LONG last_key_time = 0;
   GetKeyboardState(kbd);
   inline bool is_key_down(uchar vk) { return kbd[vk] & 0x80; }
 #ifdef debug_virtual_key_codes
-  printf(" [%d %c%d] Shift %d:%d/%d Ctrl %d:%d/%d Alt %d:%d/%d\n",
+  printf("-- [%u %c%u] Shift %d:%d/%d Ctrl %d:%d/%d Alt %d:%d/%d\n",
          (int)GetMessageTime(), lctrl_time ? '+' : '=', (int)GetMessageTime() - lctrl_time,
          is_key_down(VK_SHIFT), is_key_down(VK_LSHIFT), is_key_down(VK_RSHIFT),
          is_key_down(VK_CONTROL), is_key_down(VK_LCONTROL), is_key_down(VK_RCONTROL),
@@ -1946,14 +1946,17 @@ static LONG last_key_time = 0;
   if (key == VK_CONTROL && !extended) {
     lctrl = true;
     lctrl_time = GetMessageTime();
+    //printf("lctrl (true) %d (%d)\n", lctrl, is_key_down(VK_LCONTROL));
   }
   else if (lctrl_time) {
     lctrl = !(key == VK_MENU && extended 
               && GetMessageTime() - lctrl_time <= cfg.ctrl_alt_delay_altgr);
     lctrl_time = 0;
+    //printf("lctrl (time) %d (%d)\n", lctrl, is_key_down(VK_LCONTROL));
   }
   else {
     lctrl = is_key_down(VK_LCONTROL) && (lctrl || !is_key_down(VK_RMENU));
+    //printf("lctrl (else) %d (%d)\n", lctrl, is_key_down(VK_LCONTROL));
   }
 
   bool numlock = kbd[VK_NUMLOCK] & 1;
@@ -1965,7 +1968,12 @@ static LONG last_key_time = 0;
   bool rctrl = is_key_down(VK_RCONTROL);
   bool ctrl = lctrl | rctrl;
   bool ctrl_lalt_altgr = cfg.ctrl_alt_is_altgr & ctrl & lalt & !ralt;
-  bool altgr0 = ralt | ctrl_lalt_altgr;
+  //bool altgr0 = ralt | ctrl_lalt_altgr;
+  // Alt/AltGr detection and handling could do with a complete revision 
+  // from scratch; on the other hand, no unnecessary risk should be taken, 
+  // so another hack is added.
+  bool lctrl0 = is_key_down(VK_LCONTROL);
+  bool altgr0 = (ralt & lctrl0) | ctrl_lalt_altgr;
 
   bool external_hotkey = false;
   if (ralt && !scancode && cfg.external_hotkeys) {
@@ -1978,6 +1986,11 @@ static LONG last_key_time = 0;
   }
 
   bool altgr = ralt | ctrl_lalt_altgr;
+  // While this should more properly reflect the AltGr modifier state, 
+  // with the current implementation it has the opposite effect;
+  // it spoils Ctrl+AltGr with modify_other_keys mode.
+  //altgr = (ralt & lctrl0) | ctrl_lalt_altgr;
+
   bool win = (is_key_down(VK_LWIN) && key != VK_LWIN)
           || (is_key_down(VK_RWIN) && key != VK_RWIN);
   trace_alt("alt %d lalt %d ralt %d altgr %d\n", alt, lalt, ralt, altgr);
@@ -2719,8 +2732,8 @@ static struct {
   }
 
   bool char_key(void) {
-    trace_alt("char_key alt %d -> %d\n", alt, lalt & !ctrl_lalt_altgr);
     alt = lalt & !ctrl_lalt_altgr;
+    trace_alt("char_key alt %d (l %d r %d altgr %d)\n", alt, lalt, ralt, altgr);
 
     // Sync keyboard layout with our idea of AltGr.
     kbd[VK_CONTROL] = altgr ? 0x80 : 0;
@@ -2733,6 +2746,10 @@ static struct {
     // Try the layout.
     if (layout())
       return true;
+
+    // This prevents AltGr from behaving like Alt in modify_other_keys mode.
+    if (altgr0)
+      return false;
 
     if (ralt) {
       // Try with RightAlt/AltGr key treated as Alt.
@@ -2972,7 +2989,7 @@ static struct {
       bool check_menu = key == VK_SPACE && !term.shortcut_override
                         && cfg.window_shortcuts && alt && !altgr && !ctrl;
 #ifdef debug_key
-      printf("mods %d (modf %d comp %d)\n", mods, term.modify_other_keys, comp_state);
+      printf("-- mods %d alt %d altgr %d/%d ctrl %d lctrl %d/%d (modf %d comp %d)\n", mods, alt, altgr, altgr0, ctrl, lctrl, lctrl0, term.modify_other_keys, comp_state);
 #endif
       if (allow_shortcut && check_menu) {
         send_syscommand(SC_KEYMENU);
@@ -2980,7 +2997,8 @@ static struct {
       }
       else if (altgr_key())
         trace_key("altgr");
-      else if (altgr && !term.modify_other_keys)
+      else if (altgr0 && !term.modify_other_keys)
+        // prevent AltGr from behaving like Alt
         trace_key("!altgr");
       else if (key != ' ' && alt_code_key(key - 'A' + 0xA))
         trace_key("alt");